+++ libs/framework-core/source/ftrack_framework_core/client/__init__.py
@@ -13,7 +13,7 @@ import ftrack_constants.framework as constants
 
 from ftrack_framework_core.client.host_connection import HostConnection
 
-from ftrack_utils.decorators import track_framework_usage
+from ftrack_utils.decorators import track_framework_usage, run_in_main_thread
 
 from ftrack_utils.framework.config.tool import get_tool_config_by_name
 
@@ -180,9 +180,7 @@ class Client(object):
         return self._tool_config_options
 
     def __init__(
-        self,
-        event_manager,
-        registry,
+        self, event_manager, registry, run_in_main_thread_wrapper=None
     ):
         '''
         Initialise Client with instance of
@@ -206,6 +204,9 @@ class Client(object):
         self.__instanced_dialogs = {}
         self._dialog = None
         self._tool_config_options = defaultdict(defaultdict)
+
+        # Set up the run_in_main_thread decorator
+        self.run_in_main_thread_wrapper = run_in_main_thread_wrapper
 
         self.logger.debug('Initialising Client {}'.format(self))
 
@@ -267,6 +268,7 @@ class Client(object):
         self.event_manager.publish.client_signal_host_changed(self.id)
 
     # Context
+    @run_in_main_thread
     def _host_context_changed_callback(self, event):
         '''Set the new context ID based on data provided in *event*'''
         # Feed the new context to the client
@@ -302,6 +304,7 @@ class Client(object):
         )
 
     # Plugin
+    @run_in_main_thread
     def on_log_item_added_callback(self, event):
         '''
         Called when a log item has added in the host.
@@ -321,6 +324,7 @@ class Client(object):
             self.id, event['data']['log_item']
         )
 
+    @run_in_main_thread
     def on_ui_hook_callback(self, event):
         '''
         Called ui_hook has been executed on host and needs to notify UI with