+++ libs/framework-core/release_notes.md
@@ -1,9 +1,12 @@
 # ftrack Framework Core library release Notes
 
+
 ## upcoming
 
+* [new] Client, Host; Run in main thread decorator used to run methods in the main thread.
 * [fix] Engine; Check enabled/disabled plugins.
 * [change] Client, Dialog; Support set_tool_config_option for any item in the tool_conifg.
+
 
 ## v2.4.0
 2024-06-04