+++ libs/utils/release_notes.md
@@ -3,7 +3,9 @@
 
 ## upcoming
 
+* [new] decorators; run_in_main_thread decorator added.
 * [changed] get_temp_path; Support temp directories.
+
 
 ## v2.3.0
 2024-06-04